requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
httpx==0.27.0
pydantic==2.9.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
tldextract==5.1.2
